Split server handler into 2 parts: handshake only handler, server channel handler. This makes handshake only handler simpler, as does not contain unrelated state
