+++ netty-websocket-http2/src/test/java/com/jauntsdn/netty/handler/codec/http2/websocketx/TerminationTest.java
@@ -31,12 +31,10 @@ import java.util.stream.Stream;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
-@Disabled
 public class TerminationTest extends AbstractTest {
   private Channel server;
   private Channel client;
